third_party/blink/web_tests/external/wpt/dom/idlharness.window_exclude=Node-expected.txt

This is a testharness.js-based test.
Found 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] ShadowRoot interface: attribute serializable
  assert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function set serializable() { [native code] }"
[FAIL] Window interface: attribute event
  assert_true: property should be enumerable expected true got false
Harness: the test ran to completion.
